MIG can be built be test not working
