+++ stock_barcodes/wizard/stock_barcodes_read_inventory_views.xml
@@ -14,7 +14,10 @@
             <group name="quant_package" position="replace">
                 <div class="mt4">
                     <strong class="d-none d-sm-block">Package</strong>
-                    <span class="fa fa-inbox d-sm-none oe_span_small_icon" />
+                    <span
+                        class="fa fa-inbox d-sm-none oe_span_small_icon"
+                        title="package"
+                    />
                     <field
                         name="package_id"
                         options="{'no_create': True}"
@@ -67,6 +70,7 @@
                                 type="object"
                                 class="btn"
                                 icon="fa-recycle"
+                                title="Remove quant"
                                 context="{'wiz_barcode_id': parent.id}"
                             />
                         </tree>
