+++ stock_barcodes/wizard/stock_barcodes_read_views.xml
@@ -232,6 +232,7 @@
                                         name="action_clean_product"
                                         type="object"
                                         icon="fa-trash fa-1x"
+                                        title="Clean product"
                                         attrs="{'invisible': [('product_id', '=', False)]}"
                                     />
                                 </div>
@@ -268,6 +269,7 @@
                                         type="object"
                                         icon="fa-trash fa-1x"
                                         attrs="{'invisible': [('lot_id', '=', False), ('lot_name', '=', False)]}"
+                                        title="Clean lot"
                                     />
                                     <button
                                         id="btn_create_lot"
@@ -275,6 +277,7 @@
                                         type="action"
                                         name="%(action_stock_barcodes_new_lot)d"
                                         icon="fa-plus fa-1x"
+                                        title="Create lot"
                                         context="{'default_product_id': product_id}"
                                         attrs="{'invisible': [('create_lot', '=', False)]}"
                                         help="Create new lot"
@@ -308,6 +311,35 @@
                                 col="1"
                             >
                                 <div class="row mt8" name="option_qty_header">
+                                    <div
+                                        class="col text-center pr-0"
+                                        groups="product.group_stock_packaging"
+                                    >
+                                        <field
+                                            name="product_packaging_ids"
+                                            invisible="1"
+                                        />
+                                        <field
+                                            name="packaging_id"
+                                            options="{'no_open': True, 'no_create': True}"
+                                            domain="[('product_id', '=', product_id)]"
+                                            force_save="1"
+                                            placeholder="Packaging"
+                                            attrs="{'invisible': [('product_packaging_ids', '=', [])]}"
+                                        />
+                                    </div>
+                                    <div
+                                        class="col text-center"
+                                        groups="product.group_stock_packaging"
+                                    >
+                                        <field
+                                            name="packaging_qty"
+                                            force_save="1"
+                                            widget="numeric_step"
+                                            options="{'auto_select': True}"
+                                            attrs="{'invisible': [('product_packaging_ids', '=', [])]}"
+                                        />
+                                    </div>
                                     <div
                                         class="col text-center pr-0"
                                         name="total_qty_header"
@@ -442,6 +474,7 @@
                             name="action_force_done"
                             type="object"
                             icon="fa-check"
+                            title="Force done"
                             attrs="{'invisible': [('visible_force_done', '=', False)]}"
                             class="btn-danger ml-auto oe_kanban_action_button btn-sm"
                             data-hotkey="8"
@@ -467,38 +500,6 @@
                     </div>
                 </div>
             </form>
-        </field>
-    </record>
-    <record id="view_stock_barcodes_read_packaging_form" model="ir.ui.view">
-        <field name="name">stock.barcodes.read.packaging.form</field>
-        <field name="model">wiz.stock.barcodes.read</field>
-        <field name="inherit_id" ref="stock_barcodes.view_stock_barcodes_read_form" />
-        <field name="groups_id" eval="[(4, ref('product.group_stock_packaging'))]" />
-        <field name="arch" type="xml">
-            <div name="total_qty_header" position="before">
-                <div class="col text-center pr-0">
-                    <field name="product_packaging_ids" invisible="1" />
-                    <field
-                        name="packaging_id"
-                        options="{'no_open': True, 'no_create': True}"
-                        domain="[('product_id', '=', product_id)]"
-                        force_save="1"
-                        placeholder="Packaging"
-                        attrs="{'invisible': [('product_packaging_ids', '=', [])]}"
-                    />
-                </div>
-            </div>
-            <div name="total_qty_field" position="before">
-                <div class="col text-center">
-                    <field
-                        name="packaging_qty"
-                        force_save="1"
-                        widget="numeric_step"
-                        options="{'auto_select': True}"
-                        attrs="{'invisible': [('product_packaging_ids', '=', [])]}"
-                    />
-                </div>
-            </div>
         </field>
     </record>
     <!--
